+++ lib/src/theme/themes/default_theme_variant.dart
@@ -65,7 +65,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -80,7 +79,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -95,7 +93,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -114,7 +111,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
         ),
       ),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -128,7 +124,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -461,7 +456,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
   ShadToastTheme primaryToastTheme() {
     return ShadToastTheme(
       alignment: Alignment.bottomRight,
-      closeIconSrc: LucideIcons.x,
+      closeIconData: LucideIcons.x,
       titleStyle: effectiveTextTheme.muted.copyWith(
         fontWeight: FontWeight.w500,
         color: colorScheme.foreground,
@@ -484,7 +479,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
   ShadToastTheme destructiveToastTheme() {
     return ShadToastTheme(
       alignment: Alignment.bottomRight,
-      closeIconSrc: LucideIcons.x,
+      closeIconData: LucideIcons.x,
       titleStyle: effectiveTextTheme.muted.copyWith(
         fontWeight: FontWeight.w500,
         color: colorScheme.destructiveForeground,
@@ -515,7 +510,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
           width: 1,
         ),
       ),
-      iconSize: const Size.square(16),
       iconColor: colorScheme.foreground,
       titleStyle: effectiveTextTheme.p.copyWith(
         color: colorScheme.foreground,
@@ -540,7 +534,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
           width: 1,
         ),
       ),
-      iconSize: const Size.square(16),
       iconColor: colorScheme.destructive,
       titleStyle: effectiveTextTheme.p.copyWith(
         color: colorScheme.destructive,
@@ -556,7 +549,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
   @override
   ShadDialogTheme primaryDialogTheme() {
     return ShadDialogTheme(
-      closeIconSrc: LucideIcons.x,
+      closeIconData: LucideIcons.x,
       radius: const BorderRadius.all(Radius.circular(8)),
       backgroundColor: colorScheme.background,
       removeBorderRadiusWhenTiny: true,
@@ -653,7 +646,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
     const bezierCurve = Cubic(0.87, 0, 0.13, 1);
     const duration = Duration(milliseconds: 300);
     return ShadAccordionTheme(
-      iconSrc: LucideIcons.chevronDown,
+      iconData: LucideIcons.chevronDown,
       padding: const EdgeInsets.symmetric(vertical: 16),
       underlineTitleOnHover: true,
       duration: duration,
@@ -706,7 +699,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
           width: 0,
         ),
       ),
-      handleSize: const Size.square(10),
+      handleSize: 10,
       mainAxisAlignment: MainAxisAlignment.start,
       crossAxisAlignment: CrossAxisAlignment.center,
       mainAxisSize: MainAxisSize.max,
@@ -801,8 +794,8 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
             const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
         navigationButtonSize: 28,
         navigationButtonIconSize: 16,
-        backNavigationButtonSrc: LucideIcons.chevronLeft,
-        forwardNavigationButtonSrc: LucideIcons.chevronRight,
+        backNavigationButtonIconData: LucideIcons.chevronLeft,
+        forwardNavigationButtonIconData: LucideIcons.chevronRight,
         navigationButtonPadding: EdgeInsets.zero,
         navigationButtonDisabledOpacity: .5,
         decoration: ShadDecoration(
@@ -872,7 +865,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       buttonVariant: ShadButtonVariant.outline,
       width: 276,
       mainAxisAlignment: MainAxisAlignment.start,
-      iconSrc: LucideIcons.calendar,
+      iconData: LucideIcons.calendar,
     );
   }