+++ playground/lib/pages/alert.dart
@@ -18,13 +18,13 @@ class AlertPage extends StatelessWidget {
           padding: const EdgeInsets.symmetric(horizontal: 40),
           child: switch (style) {
             ShadAlertVariant.primary => const ShadAlert(
-                iconSrc: LucideIcons.terminal,
+                iconData: LucideIcons.terminal,
                 title: Text('Heads up!'),
                 description:
                     Text('You can add components to your app using the cli.'),
               ),
             ShadAlertVariant.destructive => const ShadAlert.destructive(
-                iconSrc: LucideIcons.circleAlert,
+                iconData: LucideIcons.circleAlert,
                 title: Text('Error'),
                 description:
                     Text('Your session has expired. Please log in again.'),